feat: enumerated DLCs
This is a pre-emptive change that looks ahead to the release of
https://store.steampowered.com/app/3816030/DayZ_Badlands/, slated for
2026. Given the current incrementing scheme, it seems reasonable to
      assume that this DLC will follow the same logic and be given the
flag 3, but this is speculation and would have to be tested on real
servers.

Separately, this commit tries to add more specificity to the DLC flags
by structuring them as Enums with a corresponding appid. Since DLC are
hardcoded to a specific appid, this value is invariate.

Providing the appid in the response could aid users of the module who
want to display additional information, and also hopefully structures
the enum better so that values like 0 and 2 are not simply a magic
number, but tied to some real-world UID.